FIXED: uint32 is erroneously defined to be 64 bit wide

+++ x11/x11_graphics_system.h
@@ -15,7 +15,7 @@ public:
     virtual ~X11GraphicsSystem() {}
 
     virtual Canvas *createCanvas(Drawable drawable, int width, int height) = 0;
-    virtual Icon *createIcon(int width, int height, X11Widget *parent, const uint32 *argb_data, uint32 bg_color) = 0;
+    virtual Icon *createIcon(int width, int height, X11Widget *parent, const unsigned long *argb_data, uint32 bg_color) = 0;
 };
 
 #endif
